# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "AssetCompleteRequest": {
        "properties": {
          "actual_size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Actual Size Bytes",
            "type": "integer"
          },
          "checksum_sha256": {
            "pattern": "^[a-fA-F0-9]{64}$",
            "title": "Checksum Sha256",
            "type": "string"
          }
        },
        "required": [
          "actual_size_bytes",
          "checksum_sha256"
        ],
        "title": "AssetCompleteRequest",
        "type": "object"
      },
      "AssetDownloadResponse": {
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetRead"
          },
          "download_url": {
            "title": "Download Url",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          }
        },
        "required": [
          "asset",
          "download_url",
          "expires_at"
        ],
        "title": "AssetDownloadResponse",
        "type": "object"
      },
      "AssetPresignRequest": {
        "properties": {
          "content_type": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Content Type",
            "type": "string"
          },
          "media_type": {
            "pattern": "^(still_image|video)$",
            "title": "Media Type",
            "type": "string"
          },
          "original_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Name"
          },
          "size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "media_type",
          "content_type",
          "size_bytes"
        ],
        "title": "AssetPresignRequest",
        "type": "object"
      },
      "AssetPresignResponse": {
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "max_size_bytes": {
            "title": "Max Size Bytes",
            "type": "integer"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "upload_url",
          "expires_at",
          "object_key",
          "max_size_bytes"
        ],
        "title": "AssetPresignResponse",
        "type": "object"
      },
      "AssetRead": {
        "properties": {
          "actual_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Size"
          },
          "checksum_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum Sha256"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expected_size": {
            "title": "Expected Size",
            "type": "integer"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "original_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Name"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "media_type",
          "content_type",
          "original_name",
          "expected_size",
          "actual_size",
          "checksum_sha256",
          "status",
          "expires_at",
          "created_at",
          "updated_at",
          "version"
        ],
        "title": "AssetRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Private project foundation for AccessForge",
    "title": "AccessForge API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/projects/{project_id}/assets": {
      "get": {
        "operationId": "list_assets_v1_projects__project_id__assets_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetRead"
                  },
                  "title": "Response List Assets V1 Projects  Project Id  Assets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Assets",
        "tags": [
          "assets"
        ]
      }
    },
    "/v1/projects/{project_id}/assets/presign-upload": {
      "post": {
        "operationId": "presign_asset_upload_v1_projects__project_id__assets_presign_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetPresignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetPresignResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Presign Asset Upload",
        "tags": [
          "assets"
        ]
      }
    },
    "/v1/projects/{project_id}/assets/{asset_id}/complete": {
      "post": {
        "operationId": "complete_asset_upload_v1_projects__project_id__assets__asset_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Asset Upload",
        "tags": [
          "assets"
        ]
      }
    },
    "/v1/projects/{project_id}/assets/{asset_id}/download": {
      "get": {
        "operationId": "download_asset_v1_projects__project_id__assets__asset_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDownloadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Asset",
        "tags": [
          "assets"
        ]
      }
    },
...
  }
}